Extract ImageMaid settings/install management into modules/imagemaid.py
Phase 5 of breaking up the quickstart.py monolith. Same shape as
Phase 4: a ~400-line settings/validation/command-building service
layer moves to modules/imagemaid.py.

Routes split into two groups based on what tests actually monkeypatch:

- probe-imagemaid-root, check-imagemaid-update, update-imagemaid,
  update-imagemaid-progress, tail-imagemaid-log move into
  blueprints/imagemaid_updates.py. Their tests only monkeypatch
  modules.helpers attributes directly (e.g. helpers.is_imagemaid_running),
  which works from any module since it mutates the shared module
  object — safe to relocate.

- validate-imagemaid and autosave-imagemaid stay as thin @app.route
  functions in quickstart.py, alongside start-imagemaid/stop-imagemaid/
  imagemaid-status (already process-control territory, untouched).
  Several tests monkeypatch qs_module._validate_imagemaid_settings,
  _persist_imagemaid_validation, _build_imagemaid_command, and
  _get_stored_plex_credentials_for_config as dependencies of these
  five routes specifically; moving the route bodies into the blueprint
  would make them call modules.imagemaid functions directly, bypassing
  quickstart.py's re-exported names entirely and silently breaking
  those patches. The underlying implementations still moved — only the
  thin route wrappers stayed, now calling through the re-exported
  aliases.

update-imagemaid depends on _get_active_work_blocker like update-kometa
did in the previous phase; same deferred `import quickstart` bridge.

Two small log-reading utilities (_is_logscan_gzip_path, _read_logscan_text)
had no quickstart-specific state and are used both by the moved
tail-imagemaid-log logic and ~9 not-yet-extracted logscan call sites
(plus one direct test reference), so they moved to modules/helpers.py
with the same re-export treatment.

Also removed two pieces of dead code found while reading this cluster:
an unused _read_text_tail function, and a second, never-called
definition of _quote_cli_value that exactly duplicated the one used
elsewhere in the file.

Net: quickstart.py shrinks by 764 lines.
