+++ quickstart.py
@@ -66,6 +66,7 @@ from modules.background_jobs import (
 from blueprints.validation_routes import bp as validation_routes_bp, refresh_plex_libraries
 from blueprints.asset_routes import bp as asset_routes_bp
 from blueprints.kometa_updates import bp as kometa_updates_bp
+from blueprints.imagemaid_updates import bp as imagemaid_updates_bp
 from modules.assets import build_preview_image_data as _build_preview_image_data, list_overlay_fonts
 from modules.kometa_install import (
     KOMETA_INSTALL_MODE_MANAGED,
@@ -75,6 +76,18 @@ from modules.kometa_install import (
     build_kometa_install_context as _build_kometa_install_context,
     get_kometa_settings_section as _get_kometa_settings_section,
     resolve_kometa_selection as _resolve_kometa_selection,
+)
+from modules.imagemaid import (
+    probe_imagemaid_root_state as _probe_imagemaid_root_state,
+    imagemaid_settings_to_form_payload as _imagemaid_settings_to_form_payload,
+    get_stored_plex_credentials_for_config as _get_stored_plex_credentials_for_config,
+    save_imagemaid_settings_for_config as _save_imagemaid_settings_for_config,
+    get_imagemaid_settings_section as _get_imagemaid_settings_section,
+    persist_imagemaid_validation as _persist_imagemaid_validation,
+    build_imagemaid_command_parts as _build_imagemaid_command_parts,
+    build_imagemaid_command as _build_imagemaid_command,
+    validate_imagemaid_settings as _validate_imagemaid_settings,
+    get_latest_imagemaid_log_path as _get_latest_imagemaid_log_path,
 )
 
 Request.max_form_parts = 100000  # Allow more form fields if needed
@@ -86,6 +99,8 @@ _safe_rel_path = helpers.safe_rel_path
 _resolve_user_dir = helpers.resolve_user_dir
 _retrieve_settings_for_config = persistence.retrieve_settings_for_config
 apply_validation_metadata = persistence.apply_validation_metadata
+_is_logscan_gzip_path = helpers.is_logscan_gzip_path
+_read_logscan_text = helpers.read_logscan_text
 
 ACTIVE_WORK_POLICIES = {
     "kometa_run": [
@@ -4051,6 +4066,7 @@ app = Flask(__name__)
 app.register_blueprint(validation_routes_bp)
 app.register_blueprint(asset_routes_bp)
 app.register_blueprint(kometa_updates_bp)
+app.register_blueprint(imagemaid_updates_bp)
 
 # Run version check at startup
 app.config["VERSION_CHECK"] = helpers.check_for_update()
@@ -9994,38 +10010,6 @@ def _detect_logscan_tool_from_path(path, log_dir=None):
     return "kometa"
 
 
-def _is_logscan_gzip_path(path):
-    try:
-        suffixes = [suffix.lower() for suffix in Path(path).suffixes]
-    except Exception:
-        return False
-    return bool(suffixes and suffixes[-1] == ".gz")
-
-
-def _read_logscan_text(path, encoding="utf-8", errors="replace"):
-    path = Path(path)
-    if _is_logscan_gzip_path(path):
-        with gzip.open(path, "rt", encoding=encoding, errors=errors) as handle:
-            return handle.read()
-    content = path.read_text(encoding=encoding, errors=errors)
-    try:
-        if path.name.lower() == "meta.log":
-            sidecar_path = path.parent / "meta.quickstart-maintenance.log"
-            if sidecar_path.exists() and sidecar_path.is_file():
-                sidecar_content = sidecar_path.read_text(encoding=encoding, errors=errors).strip()
-                if sidecar_content:
-                    content = f"{content.rstrip()}\n{sidecar_content}\n"
-        elif path.suffix.lower() == ".log":
-            sidecar_path = path.parent / "imagemaid.quickstart-maintenance.log"
-            if sidecar_path.exists() and sidecar_path.is_file():
-                sidecar_content = sidecar_path.read_text(encoding=encoding, errors=errors).strip()
-                if sidecar_content:
-                    content = f"{content.rstrip()}\n{sidecar_content}\n"
-    except Exception:
-        pass
-    return content
-
-
 def _iter_logscan_text_lines(path, encoding="utf-8", errors="replace"):
     path = Path(path)
     if _is_logscan_gzip_path(path):
@@ -14388,756 +14372,6 @@ def header_style_previews():
     return jsonify(success=True, previews=previews)
 
 
-def _probe_imagemaid_root_state(path_obj):
-    p = Path(path_obj)
-    imagemaid_root_posix = p.as_posix()
-    imagemaid_root_display = str(p)
-    is_windows = sys.platform.startswith("win")
-    venv_dir = p / "imagemaid-venv"
-    bin_dir = venv_dir / ("Scripts" if is_windows else "bin")
-    python_bin = bin_dir / ("python.exe" if is_windows else "python3")
-    if not python_bin.exists():
-        python_bin = bin_dir / ("python.exe" if is_windows else "python")
-    imagemaid_py = p / "imagemaid.py"
-    requirements = p / "requirements.txt"
-    config_dir = p / "config"
-    capabilities = _get_imagemaid_supported_options(p)
-
-    return {
-        "imagemaid_root": imagemaid_root_posix,
-        "imagemaid_root_display": imagemaid_root_display,
-        "venv_python": python_bin.as_posix(),
-        "venv_python_display": str(python_bin),
-        "root_exists": p.exists(),
-        "config_dir_exists": config_dir.exists(),
-        "imagemaid_installed": imagemaid_py.exists() and requirements.exists(),
-        "venv_exists": venv_dir.exists(),
-        "venv_python_exists": python_bin.exists(),
-        "imagemaid_running": helpers.is_imagemaid_running(),
-        "local_version": helpers.get_imagemaid_local_version(p) or "",
-        "local_sha": helpers.get_imagemaid_local_sha(p) or "",
-        "local_branch": helpers.get_imagemaid_local_branch(p) or "",
-        "supports_no_verify_ssl": bool(capabilities.get("no_verify_ssl")),
-        "supports_overlays_only": bool(capabilities.get("overlays_only")),
-    }
-
-
-def _imagemaid_settings_to_form_payload(payload):
-    if not isinstance(payload, dict):
-        return {}
-    keys = [
-        "branch_override",
-        "plex_path",
-        "mode",
-        "timeout",
-        "sleep",
-        "photo_transcoder",
-        "empty_trash",
-        "clean_bundles",
-        "optimize_db",
-        "local_db",
-        "use_existing",
-        "ignore_running",
-        "trace",
-        "log_requests",
-        "no_verify_ssl",
-        "overlays_only",
-    ]
-    form_payload = {}
-    raw_config_name = str(payload.get("config_name") or "").strip()
-    config_name = helpers.normalize_config_name_for_storage(raw_config_name) if raw_config_name else ""
-    if config_name:
-        form_payload["config_name"] = config_name
-    for key in keys:
-        if key in payload:
-            form_payload[f"imagemaid_{key}"] = payload.get(key)
-    return form_payload
-
-
-def _get_stored_plex_credentials_for_config(config_name):
-    try:
-        stored_validated, stored_user_entered, stored_payload = database.retrieve_section_data(name=config_name, section="plex")
-        payload = stored_payload if isinstance(stored_payload, dict) else {}
-        plex_settings = payload.get("plex", {})
-        if not plex_settings:
-            plex_settings = persistence.get_dummy_data("plex")
-        plex_url = plex_settings.get("url")
-        plex_token = plex_settings.get("token")
-        if plex_url and plex_token:
-            return plex_url, plex_token
-    except Exception as exc:
-        if app.config["QS_DEBUG"]:
-            helpers.ts_log(f"Failed to retrieve Plex credentials for config {config_name}: {exc}", level="ERROR")
-    return None, None
-
-
-def _save_imagemaid_settings_for_config(config_name, form_payload):
-    clean_data = persistence.clean_form_data(form_payload)
-    data = helpers.build_config_dict("imagemaid", clean_data)
-    stored_validated, stored_user_entered, stored_payload = database.retrieve_section_data(config_name, "imagemaid")
-    existing_payload = stored_payload if isinstance(stored_payload, dict) else {}
-    existing_section = existing_payload.get("imagemaid", {}) if isinstance(existing_payload.get("imagemaid"), dict) else {}
-    canonical_existing_section = _canonicalize_imagemaid_section(existing_section)
-    canonical_new_section = _canonicalize_imagemaid_section(data.get("imagemaid", {}) if isinstance(data.get("imagemaid"), dict) else {})
-    payload = dict(existing_payload)
-    payload["imagemaid"] = canonical_new_section
-    if "validated_at" in data:
-        payload["validated_at"] = data.get("validated_at")
-    elif existing_payload.get("validated_at") is not None:
-        payload["validated_at"] = existing_payload.get("validated_at")
-    user_entered = bool(canonical_new_section.get("plex_path"))
-    database.save_section_data(
-        name=config_name,
-        section="imagemaid",
-        validated=helpers.booler(stored_validated),
-        user_entered=user_entered,
-        data=payload,
-    )
-    changed = canonical_new_section != canonical_existing_section
-    return payload, changed
-
-
-def _canonicalize_imagemaid_section(section_data):
-    section = dict(section_data) if isinstance(section_data, dict) else {}
-    canonical = {
-        "branch_override": "",
-        "plex_path": "",
-        "mode": "report",
-        "timeout": 600,
-        "sleep": 60,
-        "photo_transcoder": False,
-        "empty_trash": False,
-        "clean_bundles": False,
-        "optimize_db": False,
-        "local_db": False,
-        "use_existing": False,
-        "ignore_running": False,
-        "trace": False,
-        "log_requests": False,
-        "no_verify_ssl": False,
-        "overlays_only": False,
-    }
-    canonical.update(section)
-
-    canonical["branch_override"] = helpers.normalize_imagemaid_branch_override(canonical.get("branch_override")) or ""
-    canonical["plex_path"] = str(canonical.get("plex_path") or "").strip()
-    canonical["mode"] = str(canonical.get("mode") or "report").strip().lower() or "report"
-
-    for numeric_key, default_value in (("timeout", 600), ("sleep", 60)):
-        raw_value = canonical.get(numeric_key)
-        try:
-            canonical[numeric_key] = int(raw_value)
-        except (TypeError, ValueError):
-            canonical[numeric_key] = default_value
-
-    for bool_key in (
-        "photo_transcoder",
-        "empty_trash",
-        "clean_bundles",
-        "optimize_db",
-        "local_db",
-        "use_existing",
-        "ignore_running",
-        "trace",
-        "log_requests",
-        "no_verify_ssl",
-        "overlays_only",
-    ):
-        canonical[bool_key] = helpers.booler(canonical.get(bool_key))
-
-    return canonical
-
-
-def _get_imagemaid_settings_section(config_name=None):
-    resolved_config = config_name or persistence.ensure_session_config_name()
-    settings = _retrieve_settings_for_config(resolved_config, "915-imagemaid") or {}
-    section = settings.get("imagemaid", {}) if isinstance(settings, dict) else {}
-    if not isinstance(section, dict):
-        section = {}
-    return settings, section
-
-
-def _persist_imagemaid_validation(config_name, section_data, is_valid, reason=None, details=None):
-    stored_validated, user_entered, stored_payload = database.retrieve_section_data(config_name, "imagemaid")
-    payload = stored_payload if isinstance(stored_payload, dict) else {}
-    payload["imagemaid"] = section_data if isinstance(section_data, dict) else {}
-    if is_valid:
-        payload["validated_at"] = utc_now_iso()
-        payload["validation_status"] = "validated"
-        payload.pop("validation_reason", None)
-        payload.pop("validation_details", None)
-        payload["validation_updated_at"] = utc_now_iso()
-        database.save_section_data(
-            name=config_name,
-            section="imagemaid",
-            validated=True,
-            user_entered=True,
-            data=payload,
-        )
-        return payload
-
-    existing_validated_at = payload.get("validated_at")
-    if existing_validated_at:
-        payload["validated_at"] = existing_validated_at
-    payload["validation_status"] = "failed"
-    payload["validation_reason"] = reason
-    payload["validation_details"] = details
-    payload["validation_updated_at"] = utc_now_iso()
-    database.save_section_data(
-        name=config_name,
-        section="imagemaid",
-        validated=False,
-        user_entered=True,
-        data=payload,
-    )
-    return payload
-
-
-def _get_imagemaid_supported_options(imagemaid_root=None):
-    root = Path(imagemaid_root or helpers.get_imagemaid_root_path())
-    script_path = root / "imagemaid.py"
-    try:
-        text = script_path.read_text(encoding="utf-8", errors="replace")
-    except Exception:
-        return {"no_verify_ssl": False, "overlays_only": False}
-
-    lowered = text.lower()
-    return {
-        "no_verify_ssl": ('"env": "no_verify_ssl"' in lowered) or ('"key": "no-verify-ssl"' in lowered) or ("--no-verify-ssl" in lowered),
-        "overlays_only": ('"env": "overlays_only"' in lowered) or ('"key": "overlays-only"' in lowered) or ("--overlays-only" in lowered),
-    }
-
-
-def _validate_imagemaid_plex_path(path_value, require_transcoder=False):
-    resolved = _resolve_user_dir(path_value)
-    required_label = "Plex path must point to the Plex config directory containing Cache, Metadata, and Plug-in Support."
-    if not resolved:
-        return False, "invalid_paths", f"{required_label} Use an absolute path."
-    if not resolved.exists() or not resolved.is_dir():
-        return False, "invalid_paths", f"{required_label} The selected path does not exist or is not a directory."
-
-    required_dirs = ["Metadata", "Plug-in Support"]
-    missing = [name for name in required_dirs if not (resolved / name).exists()]
-    if missing:
-        return False, "invalid_paths", f"{required_label} The selected folder is missing: {', '.join(missing)}."
-
-    if require_transcoder and not (resolved / "Cache" / "PhotoTranscoder").exists():
-        return False, "invalid_paths", f"{required_label} PhotoTranscoder cleanup also requires Cache\\PhotoTranscoder."
-
-    return True, None, None
-
-
-def _quote_cli_value(value):
-    text = str(value or "")
-    escaped = text.replace('"', '\\"')
-    return f'"{escaped}"'
-
-
-def _build_imagemaid_command_parts(section_data, plex_url, plex_token, imagemaid_root=None, redact=False):
-    imagemaid_root = Path(imagemaid_root or helpers.get_imagemaid_root_path())
-    capabilities = _get_imagemaid_supported_options(imagemaid_root)
-    is_win = sys.platform.startswith("win")
-    python_bin = imagemaid_root / "imagemaid-venv" / ("Scripts" if is_win else "bin") / ("python.exe" if is_win else "python3")
-    if not python_bin.exists():
-        python_bin = imagemaid_root / "imagemaid-venv" / ("Scripts" if is_win else "bin") / ("python.exe" if is_win else "python")
-    script_path = imagemaid_root / "imagemaid.py"
-
-    parts = [str(python_bin), str(script_path)]
-    parts.extend(["--url", "(saved Plex URL)" if redact else str(plex_url or "")])
-    parts.extend(["--token", "(saved Plex token)" if redact else str(plex_token or "")])
-
-    plex_path = str(section_data.get("plex_path") or "").strip()
-    if plex_path:
-        parts.extend(["--plex", plex_path])
-
-    mode = str(section_data.get("mode") or "report").strip().lower() or "report"
-    parts.extend(["--mode", mode])
-
-    bool_flags = {
-        "photo_transcoder": "--photo-transcoder",
-        "empty_trash": "--empty-trash",
-        "clean_bundles": "--clean-bundles",
-        "optimize_db": "--optimize-db",
-        "local_db": "--local",
-        "use_existing": "--existing",
-        "ignore_running": "--ignore",
-        "trace": "--trace",
-        "log_requests": "--log-requests",
-    }
-    if capabilities.get("no_verify_ssl"):
-        bool_flags["no_verify_ssl"] = "--no-verify-ssl"
-    if capabilities.get("overlays_only"):
-        bool_flags["overlays_only"] = "--overlays-only"
-    for key, flag in bool_flags.items():
-        if helpers.booler(section_data.get(key)):
-            parts.append(flag)
-
-    timeout_value = str(section_data.get("timeout") or "").strip()
-    if timeout_value:
-        parts.extend(["--timeout", timeout_value])
-    sleep_value = str(section_data.get("sleep") or "").strip()
-    if sleep_value:
-        parts.extend(["--sleep", sleep_value])
-
-    return parts
-
-
-def _build_imagemaid_command(section_data, plex_url, plex_token, imagemaid_root=None, redact=False):
-    parts = _build_imagemaid_command_parts(
-        section_data,
-        plex_url,
-        plex_token,
-        imagemaid_root=imagemaid_root,
-        redact=redact,
-    )
-    return subprocess.list2cmdline(parts)
-
-
-def _validate_imagemaid_settings(section_data, config_name=None):
-    if config_name:
-        resolved_config = config_name
-    elif has_request_context():
-        resolved_config = session.get("config_name") or persistence.ensure_session_config_name()
-    else:
-        resolved_config = None
-    mode = str(section_data.get("mode") or "report").strip().lower() or "report"
-    valid_modes = {"report", "move", "restore", "clear", "remove", "nothing"}
-    if mode not in valid_modes:
-        return False, "invalid_mode", f"ImageMaid mode must be one of: {', '.join(sorted(valid_modes))}."
-
-    plex_settings = _retrieve_settings_for_config(resolved_config, "010-plex") if resolved_config else (persistence.retrieve_settings("010-plex") or {})
-    if not helpers.booler(plex_settings.get("validated", False)):
-        return False, "missing_plex_validation", "Validate the Plex page before running ImageMaid."
-    if resolved_config:
-        plex_url, plex_token = _get_stored_plex_credentials_for_config(resolved_config)
-    else:
-        plex_url, plex_token = persistence.get_stored_plex_credentials()
-    if not plex_url or not plex_token:
-        return False, "missing_credentials", "Saved Plex URL/token are required."
-    valid_path, reason, details = _validate_imagemaid_plex_path(
-        section_data.get("plex_path"),
-        require_transcoder=helpers.booler(section_data.get("photo_transcoder")),
-    )
-    if not valid_path:
-        return False, reason, details
-
-    resolved_plex_path = _resolve_user_dir(section_data.get("plex_path"))
-    restore_dir = resolved_plex_path / "ImageMaid Restore" if resolved_plex_path else None
-    if mode in {"report", "move", "remove"} and restore_dir and restore_dir.exists():
-        return (
-            False,
-            "restore_dir_blocks_mode",
-            f"{mode.capitalize()} mode is not allowed while the ImageMaid Restore folder exists: {restore_dir}. " "Use nothing, restore, or clear while that folder is present.",
-        )
-    if mode in {"restore", "clear"} and restore_dir and not restore_dir.exists():
-        return False, "missing_restore_dir", f"{mode.capitalize()} mode expects the ImageMaid Restore folder at: {restore_dir}"
-
-    return True, None, None
-
-
-def _get_latest_imagemaid_log_path():
-    log_dir = helpers.get_imagemaid_root_path() / "config" / "logs"
-    if not log_dir.exists():
-        return None
-    candidates = sorted(
-        [path for path in log_dir.glob("*.log") if path.name.lower() != "imagemaid.quickstart-maintenance.log"],
-        key=lambda path: path.stat().st_mtime,
-        reverse=True,
-    )
-    return candidates[0] if candidates else None
-
-
-def _read_text_tail(path, max_lines=200):
-    content = _read_logscan_text(path, encoding="utf-8", errors="replace")
-    return "".join(deque(content.splitlines(keepends=True), maxlen=max_lines))
-
-
-def _read_text_tail_payload(path, max_lines=200):
-    path_obj = Path(path)
-    content = _read_logscan_text(path_obj, encoding="utf-8", errors="replace")
-    lines = content.splitlines()
-    tail_text = "\n".join(lines[-max_lines:]) if lines else ""
-    if tail_text and content.endswith("\n"):
-        tail_text += "\n"
-    try:
-        stats = path_obj.stat()
-        log_age_seconds = max(0, int(time.time() - stats.st_mtime))
-    except Exception:
-        log_age_seconds = None
-    return {
-        "text": tail_text,
-        "total_lines": len(lines),
-        "log_age_seconds": log_age_seconds,
-    }
-
-
-def _sanitize_imagemaid_log_tail(text):
-    content = str(text or "")
-    if not content:
-        return content
-    lines = content.splitlines()
-    filtered = []
-    i = 0
-    while i < len(lines):
-        line = lines[i]
-        if (
-            line.startswith("Exception ignored in: <function _ExecutorManagerThread.__init__.<locals>.weakref_cb")
-            and i + 3 < len(lines)
-            and lines[i + 1] == "Traceback (most recent call last):"
-            and "concurrent" in lines[i + 2]
-            and "process.py" in lines[i + 2]
-            and "AttributeError: 'NoneType' object has no attribute 'debug'" in lines[i + 3]
-        ):
-            i += 4
-            continue
-        filtered.append(line)
-        i += 1
-    return "\n".join(filtered).strip("\n")
-
-
-@app.route("/probe-imagemaid-root", methods=["POST"])
-def probe_imagemaid_root():
-    payload = request.get_json(silent=True) or {}
-    root_path = str(payload.get("path", "")).strip()
-    branch_override = helpers.normalize_imagemaid_branch_override(payload.get("branch_override"))
-    logs = []
-
-    def log(msg):
-        print(msg, file=sys.stderr)
-        logs.append(msg)
-
-    if not root_path:
-        root_path = str(helpers.get_imagemaid_root_path())
-
-    p = _resolve_user_dir(root_path)
-    if not p:
-        log("❌ Invalid path provided.")
-        return jsonify(success=False, error="Invalid path provided.", log=logs), 400
-
-    session["imagemaid_root"] = p.as_posix()
-    app.config["IMAGEMAID_ROOT"] = str(p)
-
-    state = _probe_imagemaid_root_state(p)
-    effective_branch = helpers.resolve_imagemaid_update_branch(branch_override)
-    log(f"🔍 Probing ImageMaid path: {state['imagemaid_root_display']}")
-    if not state["root_exists"]:
-        log("ℹ️ ImageMaid root does not exist yet. Install required.")
-    elif not state["imagemaid_installed"]:
-        log("ℹ️ ImageMaid files not found yet. Install required.")
-    else:
-        log("✅ ImageMaid files detected locally.")
-        if state["local_branch"]:
-            log(f"🌿 Local ImageMaid branch metadata: {state['local_branch']}")
-        if state["local_sha"]:
-            log(f"🔎 Local ImageMaid SHA: {state['local_sha'][:12]}")
-        if state["venv_python_exists"]:
-            log(f"🐍 ImageMaid venv python detected at: {state['venv_python_display']}")
-        else:
-            log("ℹ️ ImageMaid venv python not present yet. Prepare step still needed.")
-    if state["imagemaid_running"]:
-        log("ℹ️ ImageMaid is currently running.")
-
-    return (
-        jsonify(
-            success=True,
-            log=logs,
-            effective_branch=effective_branch,
-            branch_source_url=f"{helpers.IMAGEMAID_GITHUB_BASE_URL}/{effective_branch}",
-            zip_source_url=helpers.IMAGEMAID_GITHUB_ZIP_URL.format(branch=effective_branch),
-            **state,
-        ),
-        200,
-    )
-
-
-@app.route("/check-imagemaid-update", methods=["POST"])
-def check_imagemaid_update():
-    payload = request.get_json(silent=True) or {}
-    root_path = str(payload.get("path", "")).strip()
-    branch_override_raw = payload.get("branch_override")
-    branch_override = helpers.normalize_imagemaid_branch_override(branch_override_raw)
-    branch = helpers.resolve_imagemaid_update_branch(branch_override)
-    logs = []
-
-    def log(msg):
-        print(msg, file=sys.stderr)
-        logs.append(msg)
-
-    if not root_path:
-        root_path = str(helpers.get_imagemaid_root_path())
-
-    if branch_override_raw and not branch_override:
-        log(f"❌ Invalid ImageMaid branch override: {branch_override_raw}")
-        return jsonify(success=False, error="Invalid ImageMaid branch override.", log=logs), 400
-
-    p = _resolve_user_dir(root_path)
-    if not p:
-        log("❌ Invalid path provided.")
-        return jsonify(success=False, error="Invalid path provided.", log=logs), 400
-
-    session["imagemaid_root"] = p.as_posix()
-    app.config["IMAGEMAID_ROOT"] = str(p)
-
-    state = _probe_imagemaid_root_state(p)
-    if not state["imagemaid_installed"]:
-        log("ℹ️ ImageMaid is not installed yet; update check skipped.")
-        response = dict(state)
-        response.update(
-            success=True,
-            log=logs,
-            update_check_completed=False,
-            imagemaid_update_available=False,
-            imagemaid_update_check_skipped=False,
-            cached=False,
-            effective_branch=branch,
-            remote_version="",
-            remote_sha="",
-            branch_mismatch=False,
-            branch_source_url=f"{helpers.IMAGEMAID_GITHUB_BASE_URL}/{branch}",
-            zip_source_url=helpers.IMAGEMAID_GITHUB_ZIP_URL.format(branch=branch),
-        )
-        return jsonify(response), 200
-
-    if state["imagemaid_running"]:
-        log("ℹ️ ImageMaid is currently running; update check skipped.")
-        response = dict(state)
-        response.update(
-            success=True,
-            log=logs,
-            update_check_completed=True,
-            imagemaid_update_check_skipped=True,
-            imagemaid_update_available=False,
-            cached=False,
-            effective_branch=branch,
-            remote_version="",
-            remote_sha="",
-            branch_mismatch=False,
-            branch_source_url=f"{helpers.IMAGEMAID_GITHUB_BASE_URL}/{branch}",
-            zip_source_url=helpers.IMAGEMAID_GITHUB_ZIP_URL.format(branch=branch),
-        )
-        return jsonify(response), 200
-
-    if branch_override:
-        log(f"⚠️ ImageMaid branch override selected: {branch_override}")
-    else:
-        log("ℹ️ ImageMaid branch selection: auto")
-
-    update_info = helpers.get_cached_imagemaid_update(
-        p,
-        force_refresh=helpers.booler(payload.get("force", False)),
-        branch_override=branch_override,
-    )
-    remote_branch = update_info.get("branch") or "develop"
-    local_version = update_info.get("local_version") or state.get("local_version") or "unknown"
-    local_branch = update_info.get("local_branch") or "unknown"
-    local_sha = update_info.get("local_sha") or ""
-    remote_version = update_info.get("remote_version") or ""
-    remote_sha = update_info.get("remote_sha") or ""
-    log(f"🌐 Remote ImageMaid branch source: {helpers.IMAGEMAID_GITHUB_BASE_URL}/{remote_branch}")
-    if local_version:
-        log(f"ℹ️ Installed ImageMaid version: {local_version}")
-    if remote_version:
-        log(f"🌐 Remote ImageMaid version: {remote_version}")
-    log(f"ℹ️ Installed ImageMaid branch metadata: {local_branch}")
-    if local_sha:
-        log(f"🔎 Local ImageMaid SHA: {local_sha[:12]}")
-    if remote_sha:
-        log(f"🔎 Remote ImageMaid SHA: {remote_sha[:12]}")
-    if update_info.get("cached"):
-        log("ℹ️ Using cached ImageMaid update lookup.")
-    if update_info.get("update_available"):
-        if update_info.get("branch_mismatch"):
-            log(f"⚠️ Installed branch '{local_branch}' differs from selected branch '{remote_branch}'.")
-        log("⬆️ ImageMaid update available.")
-    else:
-        log("✅ ImageMaid is up to date.")
-
-    response = dict(state)
-    response.update(
-        success=True,
-        log=logs,
-        update_check_completed=True,
-        imagemaid_update_check_skipped=False,
-        imagemaid_update_available=bool(update_info.get("update_available")),
-        cached=bool(update_info.get("cached")),
-        effective_branch=remote_branch,
-        local_version=local_version,
-        local_branch=local_branch,
-        local_sha=local_sha,
-        remote_version=remote_version,
-        remote_sha=remote_sha,
-        branch_mismatch=bool(update_info.get("branch_mismatch")),
-        branch_source_url=f"{helpers.IMAGEMAID_GITHUB_BASE_URL}/{remote_branch}",
-        zip_source_url=helpers.IMAGEMAID_GITHUB_ZIP_URL.format(branch=remote_branch),
-    )
-    return jsonify(response), 200
-
-
-@app.route("/update-imagemaid", methods=["POST"])
-def update_imagemaid():
-    blocker = _get_active_work_blocker("imagemaid_update")
-    if blocker:
-        return (
-            jsonify(
-                {
-                    "success": False,
-                    "error": blocker.get("message") or "ImageMaid is currently running.",
-                    "blocked_by": blocker.get("blocked_by"),
-                    "pid": blocker.get("pid"),
-                    "target_page": blocker.get("target_page"),
-                }
-            ),
-            409,
-        )
-
-    try:
-        cfg_dir = helpers.CONFIG_DIR
-        data = request.get_json(silent=True) or {}
-        branch_override_raw = data.get("branch_override")
-        branch_override = helpers.normalize_imagemaid_branch_override(branch_override_raw)
-        if branch_override_raw and not branch_override:
-            return jsonify({"success": False, "error": "Invalid ImageMaid branch override.", "log": ["❌ Invalid ImageMaid branch override."]}), 400
-
-        imagemaid_branch = branch_override or helpers.resolve_imagemaid_update_branch()
-        force_update = helpers.booler(data.get("force", False))
-        background = data.get("background") is True
-
-        if background:
-            active = _get_active_background_job("imagemaid_update")
-            if active:
-                phase = active.get("phase")
-                if phase and phase not in ["done", "error"]:
-                    return jsonify(success=True, active=True, existing_job=True, job_id=active.get("job_id"), phase=phase), 200
-                _clear_active_background_job("imagemaid_update", job_id=active.get("job_id"))
-
-            job = _create_background_job(
-                "imagemaid_update",
-                trigger="manual",
-                phase="queued",
-                status="running",
-                target_page=JOB_TARGET_PAGES.get("imagemaid_update"),
-                logs=[],
-                done=False,
-                success=False,
-                up_to_date=False,
-                skipped=False,
-                force=force_update,
-                imagemaid_branch=imagemaid_branch,
-                started_epoch=time.time(),
-            )
-            job_id = job["job_id"]
-
-            def worker():
-                class _ProgressLog(list):
-                    def append(self_inner, item):
-                        super().append(item)
-                        current = _get_background_job(job_id) or {}
-                        lines = list(current.get("logs") or [])
-                        lines.append(item)
-                        _update_background_job(job_id, logs=lines)
-
-                logs = _ProgressLog()
-                _update_background_job(job_id, phase="running", status="running")
-                if branch_override:
-                    logs.append(f"⚠️ ImageMaid branch override selected: {branch_override}")
-                else:
-                    logs.append("ℹ️ ImageMaid branch selection: auto")
-                logs.append(f"⚙️ ImageMaid branch selected: {imagemaid_branch} (ZIP mode)")
-                if force_update:
-                    logs.append("Force update enabled.")
-
-                try:
-                    result = helpers.perform_imagemaid_update_zip_only(cfg_dir, branch=imagemaid_branch, force=force_update, logs=logs)
-                    try:
-                        helpers.invalidate_cached_imagemaid_update(Path(cfg_dir) / "imagemaid")
-                    except Exception:
-                        pass
-                    if result.get("success", False):
-                        _complete_background_job(
-                            job_id, phase="done", success=True, done=True, up_to_date=bool(result.get("up_to_date", False)), skipped=bool(result.get("skipped", False))
-                        )
-                    else:
-                        _fail_background_job(
-                            job_id,
-                            "ImageMaid update failed.",
-                            done=True,
-                            success=False,
-                            up_to_date=bool(result.get("up_to_date", False)),
-                            skipped=bool(result.get("skipped", False)),
-                        )
-                except Exception as e:
-                    logs.append("Exception during ImageMaid update.")
-                    helpers.ts_log(f"ImageMaid update failed: {e}", level="ERROR")
-                    _fail_background_job(job_id, e, done=True, success=False)
-                finally:
-                    _clear_active_background_job("imagemaid_update", job_id=job_id)
-
-            threading.Thread(target=worker, daemon=True).start()
-            return jsonify(success=True, active=True, job_id=job_id, phase="queued"), 200
-
-        logs = []
-        if branch_override:
-            logs.append(f"⚠️ ImageMaid branch override selected: {branch_override}")
-        else:
-            logs.append("ℹ️ ImageMaid branch selection: auto")
-        logs.append(f"⚙️ ImageMaid branch selected: {imagemaid_branch} (ZIP mode)")
-        if force_update:
-            logs.append("Force update enabled.")
-
-        result = helpers.perform_imagemaid_update_zip_only(cfg_dir, branch=imagemaid_branch, force=force_update, logs=logs)
-        try:
-            helpers.invalidate_cached_imagemaid_update(Path(cfg_dir) / "imagemaid")
-        except Exception:
-            pass
-        status = 200 if result.get("success") else 500
-
-        return (
-            jsonify(
-                {
-                    "success": result.get("success", False),
-                    "log": list(logs),
-                    "imagemaid_branch": imagemaid_branch,
-                    "up_to_date": result.get("up_to_date", False),
-                    "skipped": result.get("skipped", False),
-                    "force": force_update,
-                }
-            ),
-            status,
-        )
-    except Exception as e:
-        helpers.ts_log(f"ImageMaid update failed: {e}", level="ERROR")
-        return jsonify({"success": False, "log": ["Exception during ImageMaid update."]}), 500
-
-
-@app.route("/update-imagemaid-progress", methods=["GET"])
-def update_imagemaid_progress():
-    job_id = request.args.get("job_id", "").strip()
-    since = request.args.get("since", "0").strip()
-    if not job_id:
-        return jsonify(success=False, error="Missing job_id."), 400
-    info = _get_background_job(job_id)
-    if not info:
-        return jsonify(success=False, error="Unknown job_id."), 404
-    try:
-        start_idx = max(int(since or "0"), 0)
-    except ValueError:
-        start_idx = 0
-    lines = list(info.get("logs") or [])
-    return jsonify(
-        success=True,
-        job_id=job_id,
-        phase=info.get("phase"),
-        done=bool(info.get("done")) or info.get("status") in {"complete", "error"},
-        update_success=bool(info.get("success")),
-        up_to_date=bool(info.get("up_to_date")),
-        skipped=bool(info.get("skipped")),
-        force=bool(info.get("force")),
-        imagemaid_branch=info.get("imagemaid_branch"),
-        lines=lines[start_idx:],
-        next_index=len(lines),
-    )
-
-
 @app.route("/validate-imagemaid", methods=["POST"])
 def validate_imagemaid():
     payload = request.get_json(silent=True) or {}
@@ -15528,37 +14762,6 @@ def imagemaid_status():
             maintenance_window=maintenance_window,
             maintenance_paused_since=maintenance_paused_since,
         )
-
-
-@app.route("/tail-imagemaid-log", methods=["GET"])
-def tail_imagemaid_log():
-    latest_log = _get_latest_imagemaid_log_path()
-    path = latest_log if latest_log and Path(latest_log).exists() else None
-    try:
-        lines_param = str(request.args.get("lines", "2000")).strip().lower()
-        max_lines = None
-        if lines_param not in ("all", "full"):
-            max_lines = max(1, min(int(lines_param), 20000))
-    except Exception:
-        max_lines = 2000
-
-    if not path or not Path(path).exists():
-        return jsonify({"error": "No ImageMaid log found."}), 404
-    try:
-        payload = _read_text_tail_payload(path, max_lines=max_lines)
-        text = _sanitize_imagemaid_log_tail(payload.get("text"))
-        return jsonify(
-            {
-                "success": True,
-                "path": str(path),
-                "text": text,
-                "total_lines": payload.get("total_lines"),
-                "log_age_seconds": payload.get("log_age_seconds"),
-                "requested_lines": "all" if max_lines is None else max_lines,
-            }
-        )
-    except Exception as e:
-        return jsonify({"error": f"Failed to read ImageMaid log: {e}"}), 500
 
 
 def _normalize_test_libraries_path(raw_path, base_dir):
